Add analyze_rabi
Test matches matlab, but there is one outstanding question:
Why isn't the dependent variable xpts passed in with data?
In order for this work, data needs to be a function of xpts.
But xpts is only created in the function??
For now, the unit test creates data from an identical xpts.
This is still a WIP.